requirements.txt
eml-parser>=1.0.0
langgraph>=0.0.15
langchain>=0.1.0
langchain-groq>=0.0.1
requests>=2.31.0
python-dotenv>=1.0.0